libc: Add arc4random
- Added arc4random and friends to the C library so we can generate
  random numbers. This implements just the basic setup and is otherwise
  the same code as OpenBSD's libc implementation.

  There is not yet support for actually getting entropy, pending a
  kernel interface to do this.

- Updated virtual memory allocations used by libc

+++ user/lib/libc/include/sys/types.h
@@ -17,7 +17,7 @@ typedef unsigned int pid_t;
 typedef unsigned int uid_t;
 typedef unsigned int gid_t;
 
-typedef unsigned int mode_t;
+typedef unsigned long mode_t;
 
 typedef unsigned int dev_t;
 typedef unsigned int ino_t;
